api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic-settings==2.2.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
alembic==1.13.1
influxdb-client==1.41.0
